+++ controllers/pets/addPet.js
@@ -1,13 +1,14 @@
 const catchAsyncWrapper = require("../../helpers/catchAsyncWrapper");
 const httpError = require("../../helpers/httpError");
-const { Pet, addSchema } = require("../../models/notice");
+const Pet = require("../../models/pet");
+const { addSchema } = require("../../schemas");
 
 const addPet = catchAsyncWrapper(async (req, res) => {
   const { error } = addSchema.validate(req.body);
 
   if (error) {
-    const errMessage = `missing required "${error.details[0].path[0]}" field`;
-    throw httpError(400, errMessage);
+    // const errMessage = `missing required "${error.details[0].path[0]}" field`;
+    throw httpError(400, error.message);
   }
 
   const { _id: owner } = req.user;
