+++ schemas/index.js
@@ -1,5 +1,8 @@
 const emailValidator = require("../schemas/emailValidator");
+const { addSchema, sellSchema } = require("./petValidator");
 
 module.exports = {
   emailValidator,
+  addSchema,
+  sellSchema,
 };
